requirements.txt
pandas
sqlalchemy
psycopg2-binary
requests
odfpy
